main: Move callhook to own file
